+++ tmp/test_app/app/models/resource_container.rb
@@ -1,8 +1,0 @@
-class ResourceContainer < Obj
-  cms_attribute :headline, type: :string
-  cms_attribute :show_in_navigation, type: :boolean
-
-  def page
-    parent.page
-  end
-end